rbtree: assume MISRA_SANE config from zephyr
In the Zephyr implementation, there were two configurations for
rbtree: MISRA_SANE and non-MISA_SANE.

The difference between the two is that:

* MISRA_SANE avoids usage of dynamic stack allocation via alloca()
* MISRA_SANE does not implement the recursive z_rb_walk() function

For this library, we are trying to avoid any kind of dynamic memory
allocation, so these two properties are desirable. Therefore,
the source has been modified to use only the MISRA_SANE configuration.

The primary downside is that struct rbtree requires more static memory
to account for the maximum possible depth of the tree, even if the tree
never grows to that depth. On a 32-bit system, this is 59 * 5 = 295
bytes (605 bytes for 64-bit system). This could be trimmed down in
the future to allow for the user to specify the max depth at compile
time.

Signed-off-by: Nick Miller <nick.miller.83@gmail.com>