rbtree: add rb_find(), use generic compare fn
Instead of providing a "less than" function, the user must now
provide a comparision function which works like qsort (<0 if a < b,
0 if equal, >0 if a > b).

The primary motivation is to be able to implement an O(lg n) rb_find()
function, which works similarly to rb_contains(), but returns a
pointer to the found node if/when comparison function returns 0, or NULL
if it never does.

The rb_find() function is useful when trying to use the rbtree as a
key-value dictionary. Prior to this change, there was not an easy way
to search the rbtree without already having a pointer to the exact
node you're searching for (in which case you could use the existing
rb_contains() function).

Signed-off-by: Nick Miller <nick.miller.83@gmail.com>